chore: prevent error thrown when image URL cannot be composed because empty string
Fixes: https://linear.app/customerio/issue/MBL-308/error-javanetmalformedurlexception-no-protocol-on-receive-rich

Customer shared with us a push that they composed in Fly using the rich push composer, that does not have an image specified in the editor, but a "image" field gets added to the push payload with an empty string. The SDK handled when the "image" property is null, but not empty.

Customer is not experiencing a crash of the app, but they are getting stacktraces reported to them. So, this commit is a chore instead of a fix.

Testing:
* I added a new automated test. The test passes with the bug fix code, fails without it.

commit-id:9a437474